Simplify in-slot representation
In-edge slots went from 16 to 4 bytes; out-edges went from 8 to 16.
This led to some changes in the graph module as well, but only to maintain functionality.

We could also have the old { id; kind; owner; } representation in a separate type.
This type would not be stored in node types, only returned by `@property` methods.
Applying this idea to both types of edge slots could lead to a nicer user-facing API.